sandbox float resources (#284)

+++ packages/prime-sandboxes/src/prime_sandboxes/models.py
@@ -34,9 +34,9 @@ class Sandbox(BaseModel):
     name: str
     docker_image: str = Field(..., alias="dockerImage")
     start_command: Optional[str] = Field(None, alias="startCommand")
-    cpu_cores: int = Field(..., alias="cpuCores")
-    memory_gb: int = Field(..., alias="memoryGB")
-    disk_size_gb: int = Field(..., alias="diskSizeGB")
+    cpu_cores: float = Field(..., alias="cpuCores")
+    memory_gb: float = Field(..., alias="memoryGB")
+    disk_size_gb: float = Field(..., alias="diskSizeGB")
     disk_mount_path: str = Field(..., alias="diskMountPath")
     gpu_count: int = Field(..., alias="gpuCount")
     network_access: bool = Field(True, alias="networkAccess")
@@ -79,9 +79,9 @@ class CreateSandboxRequest(BaseModel):
     name: str
     docker_image: str
     start_command: Optional[str] = "tail -f /dev/null"
-    cpu_cores: int = 1
-    memory_gb: int = 2
-    disk_size_gb: int = 5
+    cpu_cores: float = 1.0
+    memory_gb: float = 2.0
+    disk_size_gb: float = 5.0
     gpu_count: int = 0
     network_access: bool = True
     timeout_minutes: int = 60
@@ -99,9 +99,9 @@ class UpdateSandboxRequest(BaseModel):
     name: Optional[str] = None
     docker_image: Optional[str] = None
     start_command: Optional[str] = None
-    cpu_cores: Optional[int] = None
-    memory_gb: Optional[int] = None
-    disk_size_gb: Optional[int] = None
+    cpu_cores: Optional[float] = None
+    memory_gb: Optional[float] = None
+    disk_size_gb: Optional[float] = None
     gpu_count: Optional[int] = None
     timeout_minutes: Optional[int] = None
     environment_vars: Optional[Dict[str, str]] = None
